feat: add EditPopup component
